refactor(trading): split shop and inventory views
