+++ bot/features/trading/views/__init__.py
@@ -1,0 +1,29 @@
+"""Trading shop and inventory Discord views."""
+
+from .common import (
+    CloseButton as CloseButton,
+)
+from .common import (
+    format_coins as format_coins,
+)
+from .inventory import (
+    InventorySelect as InventorySelect,
+)
+from .inventory import (
+    InventoryView as InventoryView,
+)
+from .shop import (
+    ShopSelect as ShopSelect,
+)
+from .shop import (
+    ShopView as ShopView,
+)
+
+__all__ = [
+    "CloseButton",
+    "InventorySelect",
+    "InventoryView",
+    "ShopSelect",
+    "ShopView",
+    "format_coins",
+]
